Add files via upload
The final assignment for the Coursera course "Python Project for Data Science". In this course, I learned how to use BeautifulSoup for web scraping and how to use yfinance to extract stock data. I learned how to work with the data obtained and how to show this data in a graph.